MAINT: Refactor to share memos
